% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctbi.plot.R
\name{ctbi.plot}
\alias{ctbi.plot}
\title{ctbi.plot}
\usage{
ctbi.plot(list.main, show.outliers = TRUE, show.n.bin = 10)
}
\arguments{
\item{list.main}{the list output from the function ctbi}

\item{show.outliers}{boolean to show or hide flagged outliers}

\item{show.n.bin}{number of bins shown within one graphic}
}
\value{
No return value
}
\description{
Plot the raw data with the bins, long-term trend and cyclic component shown.
}
