% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.check.bin.period.R
\name{hidd.check.bin.period}
\alias{hidd.check.bin.period}
\title{hidd.check.bin.period}
\usage{
hidd.check.bin.period(bin.period)
}
\arguments{
\item{bin.period}{a character string or a numeric}
}
\value{
A list that contains:

number, a numeric that indicates the value of bin.period

units, a character that indicates the unit of bin.period

bin.period.value.seconds, a numeric that indicates the value in seconds of bin.period

bin.period.value.days, a numeric that indicates the value in days of bin.period
}
\description{
interpret the string character bin.period used in ctbi.timeseries or ctbi.main
}
