% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.count.NA.R
\name{hidd.count.NA}
\alias{hidd.count.NA}
\title{hidd.count.NA}
\usage{
hidd.count.NA(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the number of NA values in a vector
}
\description{
Calculate the number of NA values in a vector
}
