% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.count.noNA.R
\name{hidd.count.noNA}
\alias{hidd.count.noNA}
\title{hidd.count.noNA}
\usage{
hidd.count.noNA(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the number of non-NA values in a vector
}
\description{
Calculate the number of non-NA values in a vector
}
