% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.mad.R
\name{hidd.mad}
\alias{hidd.mad}
\title{hidd.mad}
\usage{
hidd.mad(x, N.min.NA)
}
\arguments{
\item{x}{a numeric vector}

\item{N.min.NA}{a numeric threshold}
}
\value{
a numeric (either NA or the MAD of x)
}
\description{
Calculate the mean absolute deviation (MAD) of a vector if the number of its non-NA values is above a threshold. Otherwise, return NA.
}
