% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.mean.R
\name{hidd.mean}
\alias{hidd.mean}
\title{hidd.mean}
\usage{
hidd.mean(x, N.min.NA)
}
\arguments{
\item{x}{a numeric vector}

\item{N.min.NA}{a numeric threshold}
}
\value{
y a numeric (either NA or the mean of x)
}
\description{
Calculate the mean of a vector if the number of its non-NA values is above a threshold. Otherwise, return NA.
}
