% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.rel.time.R
\name{hidd.rel.time}
\alias{hidd.rel.time}
\title{hidd.rel.time}
\usage{
hidd.rel.time(x, seq.bin.side)
}
\arguments{
\item{x}{a vector (numeric, POSIXct or Date)}

\item{seq.bin.side}{the sequence of bin sides}
}
\value{
the relative position of each value of x with respect to the bins in seq.bin.side (between 0 and 1)
}
\description{
calculate the relative position of each timestep of a vector with respect to the bin boundaries.
}
