% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.sd.R
\name{hidd.sd}
\alias{hidd.sd}
\title{hidd.sd}
\usage{
hidd.sd(x, N.min.NA)
}
\arguments{
\item{x}{a numeric vector}

\item{N.min.NA}{a numeric threshold}
}
\value{
a numeric (either NA or the standard deviation of x)
}
\description{
Calculate the standard deviation of a vector if the number of its non-NA values is above a threshold. Otherwise, return NA.
}
