% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.seq.R
\name{hidd.seq}
\alias{hidd.seq}
\title{hidd.seq}
\usage{
hidd.seq(
  from = 1,
  to = 1,
  by = ((to - from)/(length.out - 1)),
  length.out = NULL
)
}
\arguments{
\item{from, to}{the starting and (maximal) end values of the sequence. Of length 1 unless just from is supplied as an unnamed argument.}

\item{by}{number: increment of the sequence.}

\item{length.out}{desired length of the sequence. A non-negative number, which for seq and seq.int will be rounded up if fractional.}
}
\value{
a vector of same class than from
}
\description{
similar to seq, except that when 'from' starts the 29, 30 or 31 of a month, seq2 adds 5 days to 'from', run seq, and then subtracts 5 days to the output. This is done because the function seq is not consistent when time series start at the end of the months.
}
