% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidd.sum.R
\name{hidd.sum}
\alias{hidd.sum}
\title{hidd.sum}
\usage{
hidd.sum(x, N.min.NA)
}
\arguments{
\item{x}{a numeric vector}

\item{N.min.NA}{a numeric threshold}
}
\value{
a numeric (either NA or the sum of x)
}
\description{
Calculate the sum of a vector if the number of its non-NA values is above a threshold. Otherwise, return NA.
}
